requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
aiofiles==23.2.1
python-multipart==0.0.6
jinja2==3.1.3
starlette==0.35.1
itsdangerous==2.1.2
pyngrok
python-dotenv
